Resolve #8: Add "All" option  (#10)
Add an "All" card option for the dealer card which will cause the simulator to return the expected return over all
dealer cards. This can be used in conjunction with a nonzero number of player cards to return the expected return
given a player hand regardless of the dealer's hand.

Also add a configurable Blackjack payout ratio and set this to 6/5 to match Bay 101 casino rules.